feat(ui): rename external repository components to just repository
