--- app/ui/src/Pages/NewProject/pages/Repository/Repository.module.scss
@@ -1,35 +1,0 @@
-@import 'Styles/variables';
-@import 'Styles/mixins';
-@import 'Styles/colors';
-
-.container {
-  display: flex;
-  flex-direction: column;
-  max-width: fit-content;
-
-  .repositories {
-    display: flex;
-    pointer-events: none;
-    max-height: fit-content;
-    max-width: fit-content;
-    border: 1px solid transparent;
-    transition: border ease 0.4s;
-
-    &.error {
-      border: 1px solid palette(error);
-    }
-    .repository {
-      pointer-events: auto;
-    }
-    & > div:nth-child(2) {
-      margin-left: 2 * $grid-unit;
-    }
-  }
-
-  .errorMessage {
-    @include font-small;
-    text-align: right;
-    color: palette(error);
-    margin-top: $grid-unit;
-  }
-}
