+++ app/ui/src/Pages/NewProject/pages/RepositoryDetails/RepositoryDetails.module.scss
