+++ app/ui/src/Pages/NewProject/pages/RepositoryDetails/RepositoryDetailsUtils.ts
