+++ app/ui/src/Pages/NewProject/pages/RepositoryIcon/Repository.module.scss
