+++ app/ui/src/Pages/NewProject/pages/RepositoryIcon/RepositoryIcon.tsx
@@ -25,7 +25,7 @@ type Props = {
   size?: SIZE;
   shouldAnimate?: boolean;
 };
-function RepositoryTypeComponent({ squareLocation, size = SIZE.MEDIUM, customSize, shouldAnimate = true }: Props) {
+function RepositoryIcon({ squareLocation, size = SIZE.MEDIUM, customSize, shouldAnimate = true }: Props) {
   const side = customSize ? customSize : sizePixels[size];
   return (
     <div
@@ -42,4 +42,4 @@ function RepositoryTypeComponent({ squareLocation, size = SIZE.MEDIUM, customSiz
   );
 }
 
-export default memo(RepositoryTypeComponent);
+export default memo(RepositoryIcon);
